Cleanup code and move to hook
